MyObserverクラスをserviceクラスに統合
