+++ lib/feature/analytics/analytics_service.dart
@@ -18,8 +18,28 @@ class AnalyticsService {
     await analytics.logEvent(name: eventName);
   }
 
-  Future<void> sendScreenView(Route<dynamic> route) async {
+  Future<void> _sendScreenView(Route<dynamic> route) async {
     final screenName = route.settings.name;
     await analytics.setCurrentScreen(screenName: screenName);
   }
+}
+
+final routeObserverProvider = Provider<MyRouteObserver>((ref) {
+  final service = ref.watch(analyticsServiceProvider);
+  return MyRouteObserver(service);
+});
+
+/// 既存のRouteObserverのdidPushメソッドをカスタマイズするためのクラス
+class MyRouteObserver extends RouteObserver<PageRoute<dynamic>> {
+  MyRouteObserver(this.service);
+  final AnalyticsService service;
+
+  @override
+  void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
+    super.didPush(route, previousRoute);
+    // 新しいルートがプッシュされたときに実行されるアクション
+    // Firebase Analyticsにスクリーンビュー情報を送信する
+    service._sendScreenView(route);
+  }
+
 }
